+++ src/TemperatureDataBaseWPF/Services/IService.cs
@@ -1,0 +1,7 @@
+﻿namespace TemperatureDataBaseWPF.Services
+{
+    public interface IService
+    {
+        public SerialDataAcquisition SerialHandler { get; }
+    }
+}