+++ src/TemperatureDataBaseWPF/Services/SerialDataAcquisition.cs
@@ -8,5 +8,9 @@ namespace TemperatureDataBaseWPF.Services
         {
             
         }
+        public string[] GetPortNames()
+        {
+            return SerialPort.GetPortNames();
+        }
     }
 }